Prevent aptitude from trying to nuke everything when purging obsolete packages, when the apt-get update failed
git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@3868 b52b6941-3400-464c-9f42-43200397181c

+++ rebeccablackos_files/tmp/configure_phase2.sh
@@ -168,8 +168,12 @@ apt-get dist-upgrade -y --force-yes					2>&1 |tee -a /usr/share/logs/package_ope
 apt-get --purge autoremove -y 						2>&1 |tee -a /usr/share/logs/package_operations/Installs/purge_autoremove.log
 
 #prevent packages removed from the repositories upstream to not make it in the ISOS
-aptitude purge ?obsolete -y 						2>&1 |tee -a /usr/share/logs/package_operations/Installs/purge_obsolete.log
-
+if [[ -f /var/cache/apt/pkgcache.bin ]]
+then
+  aptitude purge ?obsolete -y 						2>&1 |tee -a /usr/share/logs/package_operations/Installs/purge_obsolete.log
+else
+  echo	"Not purging older packages, because apt-get update failed"	2>&1 |tee -a /usr/share/logs/package_operations/Installs/purge_obsolete.log
+fi
 #Reset the utilites back to the way they are supposed to be.
 RevertFile /usr/sbin/grub-probe
 RevertFile /sbin/initctl
